+++ front/src/Utils/DeepPartial.ts
@@ -1,0 +1,5 @@
+export type DeepPartial<T> = T extends object
+    ? {
+          [P in keyof T]?: DeepPartial<T[P]>;
+      }
+    : T;
