+++ front/style/wa-theme/action-menu-bar.scss
@@ -17,7 +17,6 @@
           @apply tw-rounded-none tw-bg-dark-purple/95 tw-w-10 tw-h-10 tw-px-2 tw-m-0 tw-outline-none;
           @apply sm:tw-w-12 sm:tw-h-12;
 
-          &.focus, &:focus,
           &.hover, &:hover {
             @apply tw-bg-medium-purple/95
           }
@@ -36,7 +35,7 @@
 
 
       #invite-btn {
-        @apply tw-min-h-[2rem] tw-h-8 sm:tw-h-10;
+        @apply tw-min-h-[2rem] tw-h-8 sm:tw-h-10 tw-outline-none;
       }
 
       .bottom-action-button:first-child {
